+++ core/src/main/kotlin/com/maltaisn/nines/core/GameListener.kt
@@ -37,6 +37,8 @@ interface GameListener : CardGameListener {
 
     fun onReportBugClicked()
 
+    fun onViewSourceClicked()
+
     fun onRateAppClicked() = Unit
 
 }